Port 69a1179 "linux-gen: packet: compress packet offset table"

+++ platform/linux-dpdk/include/odp/api/plat/packet_types.h
@@ -64,35 +64,35 @@ typedef enum {
 /** @internal Packet header field offsets for inline functions */
 typedef struct _odp_packet_inline_offset_t {
 	/** @internal field offset */
-	size_t mb;
+	uint16_t mb;
 	/** @internal field offset */
-	size_t pool;
+	uint16_t pool;
 	/** @internal field offset */
-	size_t input;
+	uint16_t input;
 	/** @internal field offset */
-	size_t user_ptr;
+	uint16_t user_ptr;
 	/** @internal field offset */
-	size_t timestamp;
+	uint16_t timestamp;
 	/** @internal field offset */
-	size_t input_flags;
+	uint16_t input_flags;
 	/** @internal field offset */
-	size_t buf_addr;
+	uint16_t buf_addr;
 	/** @internal field offset */
-	size_t data;
+	uint16_t data;
 	/** @internal field offset */
-	size_t pkt_len;
+	uint16_t pkt_len;
 	/** @internal field offset */
-	size_t seg_len;
+	uint16_t seg_len;
 	/** @internal field offset */
-	size_t nb_segs;
+	uint16_t nb_segs;
 	/** @internal field offset */
-	size_t udata_len;
+	uint16_t udata_len;
 	/** @internal offset */
-	size_t udata;
+	uint16_t udata;
 	/** @internal field offset */
-	size_t rss;
+	uint16_t rss;
 	/** @internal field offset */
-	size_t ol_flags;
+	uint16_t ol_flags;
 	/** @internal rss hash result set */
 	uint64_t rss_flag;
